// src/globals.d.ts
namespace Globals {
    type State = {
        adsSens: number
        aspectRation: Array<number, number>
        fov: number
        customAds: Array<
            number,
            number,
            number,
            number,
            number,
            number,
            number,
            number
        >
    }
}

export { Globals }